Cleaned Convert and Utils Lib
Removed outdated an obsolute x2x blocks. Moved several struct and array
conversion blocks from utils to convert.
