+++ modules/convert/include/forte/eclipse4diac/convert/GEN_ARRAY2VALUES_fbt.h
@@ -1,6 +1,8 @@
 /*******************************************************************************
- * Copyright (c) 2014 Profactor GmbH, fortiss GmbH
- *                      2018 Johannes Kepler University
+ * Copyright (c) 2014, 2023 Profactor GmbH, fortiss GbmH
+ *                          Johannes Kepler University
+ *                          Martin Erich Jobst
+ *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
@@ -19,33 +21,38 @@ class for better handling generic FBs
 #include "forte/genfb.h"
 
 #include <memory>
+#include "forte/dataconn.h"
 #include "forte/datatypes/forte_any_variant.h"
 #include "forte/datatypes/forte_array_dynamic.h"
 
-namespace forte::eclipse4diac::utils {
-  class GEN_VALUES2ARRAY final : public CGenFunctionBlock<CFunctionBlock> {
-      DECLARE_GENERIC_FIRMWARE_FB(GEN_VALUES2ARRAY)
+namespace forte::eclipse4diac::convert {
+  class GEN_ARRAY2VALUES final : public CGenFunctionBlock<CFunctionBlock> {
+      DECLARE_GENERIC_FIRMWARE_FB(GEN_ARRAY2VALUES)
 
     protected:
+      void createGenOutputData() override;
+
       size_t getGenEOOffset() override {
         return 1;
       }
 
-      size_t getGenDOOffset() override {
+      size_t getGenDIOffset() override {
         return 1;
       }
-
-      void createGenInputData() override;
 
       CEventConnection *getEOConUnchecked(TPortId paEONum) override;
       CIEC_ANY *getDI(size_t paIndex) override;
       CIEC_ANY *getDO(size_t paIndex) override;
-      CDataConnection *getDOConUnchecked(TPortId paDONum) override;
+      CDataConnection **getDIConUnchecked(const TPortId paIndex) override;
 
     private:
+      std::vector<StringId> mDataOutputNames;
+
       static const TEventID scmEventREQID = 0;
 
       static const TEventID scmEventCNFID = 0;
+
+      // self-defined members
 
       void executeEvent(TEventID paEIID, CEventChainExecutionThread *const paECET) override;
 
@@ -54,14 +61,15 @@ namespace forte::eclipse4diac::utils {
 
       bool createInterfaceSpec(const char *paConfigString, SFBInterfaceSpec &paInterfaceSpec) override;
 
-      std::vector<StringId> mDataInputNames;
       CEventConnection conn_CNF;
-      CIEC_ARRAY_DYNAMIC var_OUT;
-      COutDataConnection<CIEC_ARRAY_DYNAMIC> conn_OUT;
-      std::unique_ptr<CIEC_ANY_VARIANT[]> mGenDIs;
+
+      CIEC_ARRAY_DYNAMIC var_IN;
+      CDataConnection *conn_IN;
+
+      std::unique_ptr<CIEC_ANY_VARIANT[]> mGenDOs;
 
     public:
-      GEN_VALUES2ARRAY(const StringId paInstanceNameId, CFBContainer &paContainer);
-      ~GEN_VALUES2ARRAY() override = default;
+      GEN_ARRAY2VALUES(const StringId paInstanceNameId, CFBContainer &paContainer);
+      ~GEN_ARRAY2VALUES() override = default;
   };
-} // namespace forte::eclipse4diac::utils
+} // namespace forte::eclipse4diac::convert
