+++ modules/convert/include/forte/eclipse4diac/convert/GEN_VALUES2ARRAY_fbt.h
@@ -1,8 +1,6 @@
 /*******************************************************************************
- * Copyright (c) 2014, 2023 Profactor GmbH, fortiss GbmH
- *                          Johannes Kepler University
- *                          Martin Erich Jobst
- *
+ * Copyright (c) 2014 Profactor GmbH, fortiss GmbH
+ *                      2018 Johannes Kepler University
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
@@ -21,38 +19,33 @@ class for better handling generic FBs
 #include "forte/genfb.h"
 
 #include <memory>
-#include "forte/dataconn.h"
 #include "forte/datatypes/forte_any_variant.h"
 #include "forte/datatypes/forte_array_dynamic.h"
 
-namespace forte::eclipse4diac::utils {
-  class GEN_ARRAY2VALUES final : public CGenFunctionBlock<CFunctionBlock> {
-      DECLARE_GENERIC_FIRMWARE_FB(GEN_ARRAY2VALUES)
+namespace forte::eclipse4diac::convert {
+  class GEN_VALUES2ARRAY final : public CGenFunctionBlock<CFunctionBlock> {
+      DECLARE_GENERIC_FIRMWARE_FB(GEN_VALUES2ARRAY)
 
     protected:
-      void createGenOutputData() override;
-
       size_t getGenEOOffset() override {
         return 1;
       }
 
-      size_t getGenDIOffset() override {
+      size_t getGenDOOffset() override {
         return 1;
       }
+
+      void createGenInputData() override;
 
       CEventConnection *getEOConUnchecked(TPortId paEONum) override;
       CIEC_ANY *getDI(size_t paIndex) override;
       CIEC_ANY *getDO(size_t paIndex) override;
-      CDataConnection **getDIConUnchecked(const TPortId paIndex) override;
+      CDataConnection *getDOConUnchecked(TPortId paDONum) override;
 
     private:
-      std::vector<StringId> mDataOutputNames;
-
       static const TEventID scmEventREQID = 0;
 
       static const TEventID scmEventCNFID = 0;
-
-      // self-defined members
 
       void executeEvent(TEventID paEIID, CEventChainExecutionThread *const paECET) override;
 
@@ -61,15 +54,14 @@ namespace forte::eclipse4diac::utils {
 
       bool createInterfaceSpec(const char *paConfigString, SFBInterfaceSpec &paInterfaceSpec) override;
 
+      std::vector<StringId> mDataInputNames;
       CEventConnection conn_CNF;
-
-      CIEC_ARRAY_DYNAMIC var_IN;
-      CDataConnection *conn_IN;
-
-      std::unique_ptr<CIEC_ANY_VARIANT[]> mGenDOs;
+      CIEC_ARRAY_DYNAMIC var_OUT;
+      COutDataConnection<CIEC_ARRAY_DYNAMIC> conn_OUT;
+      std::unique_ptr<CIEC_ANY_VARIANT[]> mGenDIs;
 
     public:
-      GEN_ARRAY2VALUES(const StringId paInstanceNameId, CFBContainer &paContainer);
-      ~GEN_ARRAY2VALUES() override = default;
+      GEN_VALUES2ARRAY(const StringId paInstanceNameId, CFBContainer &paContainer);
+      ~GEN_VALUES2ARRAY() override = default;
   };
-} // namespace forte::eclipse4diac::utils
+} // namespace forte::eclipse4diac::convert
