+++ modules/convert/include/forte/eclipse4diac/convert/GET_STRUCT_VALUE_fbt.h
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018, 2023 fortiss GmbH, Primetals Technologies Austria GmbH
+ * Copyright (c) 2018, fortiss GmbH
  *               2023 Martin Erich Jobst
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -9,7 +9,6 @@
  *
  * Contributors:
  *   Jose Cabral - initial implementation
- *   Alois Zoitl - adapted this FB from the code in GET_STRUCT_VALUE
  *   Martin Jobst
  *     - refactor for ANY variant
  *******************************************************************************/
@@ -19,15 +18,16 @@
 #include "forte/funcbloc.h"
 #include "forte/datatypes/forte_any_variant.h"
 #include "forte/datatypes/forte_string.h"
+#include "forte/datatypes/forte_bool.h"
 #include "forte/iec61131_functions.h"
 #include "forte/datatypes/forte_array_common.h"
 #include "forte/datatypes/forte_array.h"
 #include "forte/datatypes/forte_array_fixed.h"
 #include "forte/datatypes/forte_array_variable.h"
 
-namespace forte::eclipse4diac::utils {
-  class FORTE_SET_STRUCT_VALUE : public CFunctionBlock {
-      DECLARE_FIRMWARE_FB(FORTE_SET_STRUCT_VALUE)
+namespace forte::eclipse4diac::convert {
+  class FORTE_GET_STRUCT_VALUE : public CFunctionBlock {
+      DECLARE_FIRMWARE_FB(FORTE_GET_STRUCT_VALUE)
 
     private:
       static const TEventID scmEventREQID = 0;
@@ -42,18 +42,18 @@ namespace forte::eclipse4diac::utils {
       void writeOutputData(TEventID paEIID) override;
 
     public:
-      FORTE_SET_STRUCT_VALUE(const StringId paInstanceNameId, CFBContainer &paContainer);
+      FORTE_GET_STRUCT_VALUE(const StringId paInstanceNameId, CFBContainer &paContainer);
 
       CIEC_ANY_VARIANT var_in_struct;
       CIEC_STRING var_member;
-      CIEC_ANY_VARIANT var_element_value;
-      CIEC_ANY_VARIANT var_out_struct;
+      CIEC_BOOL var_QO;
+      CIEC_ANY_VARIANT var_output;
 
       CEventConnection conn_CNF;
       CDataConnection *conn_in_struct;
       CDataConnection *conn_member;
-      CDataConnection *conn_element_value;
-      COutDataConnection<CIEC_ANY_VARIANT> conn_out_struct;
+      COutDataConnection<CIEC_BOOL> conn_QO;
+      COutDataConnection<CIEC_ANY_VARIANT> conn_output;
 
       CIEC_ANY *getDI(size_t) override;
       CIEC_ANY *getDO(size_t) override;
@@ -63,24 +63,25 @@ namespace forte::eclipse4diac::utils {
 
       void evt_REQ(const CIEC_ANY &pa_in_struct,
                    const CIEC_STRING &pa_member,
-                   const CIEC_ANY &pa_element_value,
-                   COutputParameter<CIEC_ANY_VARIANT> pa_out_struct) {
-        COutputGuard guard_pa_out_struct(pa_out_struct);
+                   CAnyBitOutputParameter<CIEC_BOOL> pa_QO,
+                   COutputParameter<CIEC_ANY_VARIANT> pa_output) {
+        COutputGuard guard_pa_QO(pa_QO);
+        COutputGuard guard_pa_output(pa_output);
         var_in_struct = pa_in_struct;
         var_member = pa_member;
-        var_element_value = pa_element_value;
         receiveInputEvent(scmEventREQID, nullptr);
-        pa_out_struct->setValue(var_out_struct.unwrap());
+        *pa_QO = var_QO;
+        pa_output->setValue(var_output.unwrap());
       }
 
       void operator()(const CIEC_ANY &pa_in_struct,
                       const CIEC_STRING &pa_member,
-                      const CIEC_ANY &pa_element_value,
-                      COutputParameter<CIEC_ANY_VARIANT> pa_out_struct) {
-        evt_REQ(pa_in_struct, pa_member, pa_element_value, pa_out_struct);
+                      CAnyBitOutputParameter<CIEC_BOOL> pa_QO,
+                      COutputParameter<CIEC_ANY_VARIANT> pa_output) {
+        evt_REQ(pa_in_struct, pa_member, pa_QO, pa_output);
       }
 
     protected:
       void setInitialValues() override;
   };
-} // namespace forte::eclipse4diac::utils
+} // namespace forte::eclipse4diac::convert
