+++ modules/convert/include/forte/eclipse4diac/convert/SET_AT_INDEX_fbt.h
@@ -26,9 +26,9 @@
 #include "forte/datatypes/forte_array_fixed.h"
 #include "forte/datatypes/forte_array_variable.h"
 
-namespace forte::eclipse4diac::utils {
-  class FORTE_GET_AT_INDEX : public CFunctionBlock {
-      DECLARE_FIRMWARE_FB(FORTE_GET_AT_INDEX)
+namespace forte::eclipse4diac::convert {
+  class FORTE_SET_AT_INDEX : public CFunctionBlock {
+      DECLARE_FIRMWARE_FB(FORTE_SET_AT_INDEX)
 
     private:
       static const TEventID scmEventREQID = 0;
@@ -41,18 +41,20 @@ namespace forte::eclipse4diac::utils {
       void writeOutputData(TEventID paEIID) override;
 
     public:
-      FORTE_GET_AT_INDEX(const StringId paInstanceNameId, CFBContainer &paContainer);
+      FORTE_SET_AT_INDEX(const StringId paInstanceNameId, CFBContainer &paContainer);
 
       CIEC_ANY_VARIANT var_IN_ARRAY;
       CIEC_UINT var_INDEX;
+      CIEC_ANY_VARIANT var_VALUE;
       CIEC_BOOL var_QO;
-      CIEC_ANY_VARIANT var_OUT;
+      CIEC_ANY_VARIANT var_OUT_ARRAY;
 
       CEventConnection conn_CNF;
       CDataConnection *conn_IN_ARRAY;
       CDataConnection *conn_INDEX;
+      CDataConnection *conn_VALUE;
       COutDataConnection<CIEC_BOOL> conn_QO;
-      COutDataConnection<CIEC_ANY_VARIANT> conn_OUT;
+      COutDataConnection<CIEC_ANY_VARIANT> conn_OUT_ARRAY;
 
       CIEC_ANY *getDI(size_t) override;
       CIEC_ANY *getDO(size_t) override;
@@ -62,25 +64,28 @@ namespace forte::eclipse4diac::utils {
 
       void evt_REQ(const CIEC_ANY &pa_IN_ARRAY,
                    const CIEC_UINT &pa_INDEX,
+                   const CIEC_ANY &pa_VALUE,
                    CAnyBitOutputParameter<CIEC_BOOL> pa_QO,
-                   COutputParameter<CIEC_ANY_VARIANT> pa_OUT) {
+                   COutputParameter<CIEC_ANY_VARIANT> pa_OUT_ARRAY) {
         COutputGuard guard_pa_QO(pa_QO);
-        COutputGuard guard_pa_OUT(pa_OUT);
+        COutputGuard guard_pa_OUT_ARRAY(pa_OUT_ARRAY);
         var_IN_ARRAY = pa_IN_ARRAY;
         var_INDEX = pa_INDEX;
+        var_VALUE = pa_VALUE;
         receiveInputEvent(scmEventREQID, nullptr);
         *pa_QO = var_QO;
-        pa_OUT->setValue(var_OUT.unwrap());
+        pa_OUT_ARRAY->setValue(var_OUT_ARRAY.unwrap());
       }
 
       void operator()(const CIEC_ANY &pa_IN_ARRAY,
                       const CIEC_UINT &pa_INDEX,
+                      const CIEC_ANY &pa_VALUE,
                       CAnyBitOutputParameter<CIEC_BOOL> pa_QO,
-                      COutputParameter<CIEC_ANY_VARIANT> pa_OUT) {
-        evt_REQ(pa_IN_ARRAY, pa_INDEX, pa_QO, pa_OUT);
+                      COutputParameter<CIEC_ANY_VARIANT> pa_OUT_ARRAY) {
+        evt_REQ(pa_IN_ARRAY, pa_INDEX, pa_VALUE, pa_QO, pa_OUT_ARRAY);
       }
 
     protected:
       void setInitialValues() override;
   };
-} // namespace forte::eclipse4diac::utils
+} // namespace forte::eclipse4diac::convert
