+++ modules/convert/src/eclipse4diac/convert/CMakeLists.txt
@@ -13,21 +13,13 @@
 #############################################################################
 
 target_sources(forte-convert PRIVATE
-        BOOL2BOOL_fbt.cpp
-        BYTE2BYTE_fbt.cpp
-        DINT2DINT_fbt.cpp
-        DWORD2DWORD_fbt.cpp
+        GEN_ARRAY2ARRAY_fbt.cpp
+        GEN_ARRAY2VALUES_fbt.cpp
         GEN_STRUCT_DEMUX_fbt.cpp
         GEN_STRUCT_MUX_fbt.cpp
-        INT2INT_fbt.cpp
-        LREAL2LREAL_fbt.cpp
-        REAL2REAL_fbt.cpp
-        SINT2SINT_fbt.cpp
-        STRING2STRING_fbt.cpp
-        TIME2TIME_fbt.cpp
-        UDINT2UDINT_fbt.cpp
-        UINT2UINT_fbt.cpp
-        USINT2USINT_fbt.cpp
-        WORD2WORD_fbt.cpp
-        WSTRING2WSTRING_fbt.cpp
+        GEN_VALUES2ARRAY_fbt.cpp
+        GET_AT_INDEX_fbt.cpp
+        GET_STRUCT_VALUE_fbt.cpp
+        SET_AT_INDEX_fbt.cpp
+        SET_STRUCT_VALUE_fbt.cpp
 )
