--- modules/convert/src/eclipse4diac/convert/DWORD2DWORD_fbt.cpp
@@ -1,138 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2023 Martin Erich Jobst
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *   Martin Jobst
- *     - initial API and implementation and/or initial documentation
- *******************************************************************************/
-
-#include "forte/eclipse4diac/convert/DWORD2DWORD_fbt.h"
-
-#include "forte/datatypes/forte_dword.h"
-#include "forte/iec61131_functions.h"
-#include "forte/datatypes/forte_array_common.h"
-#include "forte/datatypes/forte_array.h"
-#include "forte/datatypes/forte_array_fixed.h"
-#include "forte/datatypes/forte_array_variable.h"
-
-using namespace forte::literals;
-
-namespace forte::eclipse4diac::convert {
-  namespace {
-    const auto cDataInputNames = std::array{"IN"_STRID};
-    const auto cDataOutputNames = std::array{"OUT"_STRID};
-    const auto cEventInputNames = std::array{"REQ"_STRID};
-    const auto cEventOutputNames = std::array{"CNF"_STRID};
-    const SFBInterfaceSpec cFBInterfaceSpec = {
-        .mEINames = cEventInputNames,
-        .mEITypeNames = {},
-        .mEONames = cEventOutputNames,
-        .mEOTypeNames = {},
-        .mDINames = cDataInputNames,
-        .mDONames = cDataOutputNames,
-        .mDIONames = {},
-        .mSocketNames = {},
-        .mPlugNames = {},
-    };
-  } // namespace
-
-  DEFINE_FIRMWARE_FB(FORTE_DWORD2DWORD, "eclipse4diac::convert::DWORD2DWORD"_STRID)
-
-  FORTE_DWORD2DWORD::FORTE_DWORD2DWORD(const StringId paInstanceNameId, CFBContainer &paContainer) :
-      CSimpleFB(paContainer, cFBInterfaceSpec, paInstanceNameId, {}),
-      var_IN(0_DWORD),
-      var_OUT(0_DWORD),
-      conn_CNF(*this, 0),
-      conn_IN(nullptr),
-      conn_OUT(*this, 0, var_OUT) {
-  }
-
-  void FORTE_DWORD2DWORD::setInitialValues() {
-    var_IN = 0_DWORD;
-    var_OUT = 0_DWORD;
-  }
-
-  void FORTE_DWORD2DWORD::executeEvent(const TEventID paEIID, CEventChainExecutionThread *const paECET) {
-    switch (paEIID) {
-      case scmEventREQID: enterStateREQ(paECET); break;
-      default: break;
-    }
-  }
-
-  void FORTE_DWORD2DWORD::enterStateREQ(CEventChainExecutionThread *const paECET) {
-    alg_REQ();
-    sendOutputEvent(scmEventCNFID, paECET);
-  }
-
-  void FORTE_DWORD2DWORD::readInputData(const TEventID paEIID) {
-    switch (paEIID) {
-      case scmEventREQID: {
-        readData(0, var_IN, conn_IN);
-        break;
-      }
-      default: break;
-    }
-  }
-
-  void FORTE_DWORD2DWORD::writeOutputData(const TEventID paEIID) {
-    switch (paEIID) {
-      case scmEventCNFID: {
-        writeData(cFBInterfaceSpec.getNumDIs() + 0, var_OUT, conn_OUT);
-        break;
-      }
-      default: break;
-    }
-  }
-
-  CIEC_ANY *FORTE_DWORD2DWORD::getDI(const size_t paIndex) {
-    switch (paIndex) {
-      case 0: return &var_IN;
-    }
-    return nullptr;
-  }
-
-  CIEC_ANY *FORTE_DWORD2DWORD::getDO(const size_t paIndex) {
-    switch (paIndex) {
-      case 0: return &var_OUT;
-    }
-    return nullptr;
-  }
-
-  CEventConnection *FORTE_DWORD2DWORD::getEOConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_CNF;
-    }
-    return nullptr;
-  }
-
-  CDataConnection **FORTE_DWORD2DWORD::getDIConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_IN;
-    }
-    return nullptr;
-  }
-
-  CDataConnection *FORTE_DWORD2DWORD::getDOConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_OUT;
-    }
-    return nullptr;
-  }
-
-  CIEC_ANY *FORTE_DWORD2DWORD::getVarInternal(size_t) {
-    return nullptr;
-  }
-
-  void FORTE_DWORD2DWORD::alg_REQ(void) {
-
-#line 2 "DWORD2DWORD.fbt"
-    var_OUT = var_IN;
-  }
-
-} // namespace forte::eclipse4diac::convert
