+++ modules/convert/src/eclipse4diac/convert/GET_AT_INDEX_fbt.cpp
@@ -14,12 +14,12 @@
  *     - refactor for ANY variant
  *******************************************************************************/
 
-#include "forte/eclipse4diac/utils/GET_AT_INDEX_fbt.h"
+#include "forte/eclipse4diac/convert/GET_AT_INDEX_fbt.h"
 #include "forte/datatypes/forte_bool.h"
 
 using namespace forte::literals;
 
-namespace forte::eclipse4diac::utils {
+namespace forte::eclipse4diac::convert {
   namespace {
     const auto cDataInputNames = std::array{"IN_ARRAY"_STRID, "INDEX"_STRID};
 
@@ -44,7 +44,7 @@ namespace forte::eclipse4diac::utils {
     };
   } // namespace
 
-  DEFINE_FIRMWARE_FB(FORTE_GET_AT_INDEX, "eclipse4diac::utils::GET_AT_INDEX"_STRID)
+  DEFINE_FIRMWARE_FB(FORTE_GET_AT_INDEX, "eclipse4diac::convert::GET_AT_INDEX"_STRID)
 
   FORTE_GET_AT_INDEX::FORTE_GET_AT_INDEX(const StringId paInstanceNameId, CFBContainer &paContainer) :
       CFunctionBlock(paContainer, cFBInterfaceSpec, paInstanceNameId),
@@ -147,4 +147,4 @@ namespace forte::eclipse4diac::utils {
     var_OUT.reset();
   }
 
-} // namespace forte::eclipse4diac::utils
+} // namespace forte::eclipse4diac::convert
