+++ modules/convert/src/eclipse4diac/convert/GET_STRUCT_VALUE_fbt.cpp
@@ -14,11 +14,11 @@
  *     - refactor for ANY variant
  *******************************************************************************/
 
-#include "forte/eclipse4diac/utils/GET_STRUCT_VALUE_fbt.h"
+#include "forte/eclipse4diac/convert/GET_STRUCT_VALUE_fbt.h"
 
 using namespace forte::literals;
 
-namespace forte::eclipse4diac::utils {
+namespace forte::eclipse4diac::convert {
   namespace {
     const auto cDataInputNames = std::array{"in_struct"_STRID, "member"_STRID};
 
@@ -43,7 +43,7 @@ namespace forte::eclipse4diac::utils {
     };
   } // namespace
 
-  DEFINE_FIRMWARE_FB(FORTE_GET_STRUCT_VALUE, "eclipse4diac::utils::GET_STRUCT_VALUE"_STRID)
+  DEFINE_FIRMWARE_FB(FORTE_GET_STRUCT_VALUE, "eclipse4diac::convert::GET_STRUCT_VALUE"_STRID)
 
   FORTE_GET_STRUCT_VALUE::FORTE_GET_STRUCT_VALUE(const StringId paInstanceNameId, CFBContainer &paContainer) :
       CFunctionBlock(paContainer, cFBInterfaceSpec, paInstanceNameId),
@@ -167,4 +167,4 @@ namespace forte::eclipse4diac::utils {
     var_output.reset();
   }
 
-} // namespace forte::eclipse4diac::utils
+} // namespace forte::eclipse4diac::convert
