--- modules/convert/src/eclipse4diac/convert/SINT2SINT_fbt.cpp
@@ -1,138 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2023 Martin Erich Jobst
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *   Martin Jobst
- *     - initial API and implementation and/or initial documentation
- *******************************************************************************/
-
-#include "forte/eclipse4diac/convert/SINT2SINT_fbt.h"
-
-#include "forte/datatypes/forte_sint.h"
-#include "forte/iec61131_functions.h"
-#include "forte/datatypes/forte_array_common.h"
-#include "forte/datatypes/forte_array.h"
-#include "forte/datatypes/forte_array_fixed.h"
-#include "forte/datatypes/forte_array_variable.h"
-
-using namespace forte::literals;
-
-namespace forte::eclipse4diac::convert {
-  namespace {
-    const auto cDataInputNames = std::array{"IN"_STRID};
-    const auto cDataOutputNames = std::array{"OUT"_STRID};
-    const auto cEventInputNames = std::array{"REQ"_STRID};
-    const auto cEventOutputNames = std::array{"CNF"_STRID};
-    const SFBInterfaceSpec cFBInterfaceSpec = {
-        .mEINames = cEventInputNames,
-        .mEITypeNames = {},
-        .mEONames = cEventOutputNames,
-        .mEOTypeNames = {},
-        .mDINames = cDataInputNames,
-        .mDONames = cDataOutputNames,
-        .mDIONames = {},
-        .mSocketNames = {},
-        .mPlugNames = {},
-    };
-  } // namespace
-
-  DEFINE_FIRMWARE_FB(FORTE_SINT2SINT, "eclipse4diac::convert::SINT2SINT"_STRID)
-
-  FORTE_SINT2SINT::FORTE_SINT2SINT(const StringId paInstanceNameId, CFBContainer &paContainer) :
-      CSimpleFB(paContainer, cFBInterfaceSpec, paInstanceNameId, {}),
-      var_IN(0_SINT),
-      var_OUT(0_SINT),
-      conn_CNF(*this, 0),
-      conn_IN(nullptr),
-      conn_OUT(*this, 0, var_OUT) {
-  }
-
-  void FORTE_SINT2SINT::setInitialValues() {
-    var_IN = 0_SINT;
-    var_OUT = 0_SINT;
-  }
-
-  void FORTE_SINT2SINT::executeEvent(const TEventID paEIID, CEventChainExecutionThread *const paECET) {
-    switch (paEIID) {
-      case scmEventREQID: enterStateREQ(paECET); break;
-      default: break;
-    }
-  }
-
-  void FORTE_SINT2SINT::enterStateREQ(CEventChainExecutionThread *const paECET) {
-    alg_REQ();
-    sendOutputEvent(scmEventCNFID, paECET);
-  }
-
-  void FORTE_SINT2SINT::readInputData(const TEventID paEIID) {
-    switch (paEIID) {
-      case scmEventREQID: {
-        readData(0, var_IN, conn_IN);
-        break;
-      }
-      default: break;
-    }
-  }
-
-  void FORTE_SINT2SINT::writeOutputData(const TEventID paEIID) {
-    switch (paEIID) {
-      case scmEventCNFID: {
-        writeData(cFBInterfaceSpec.getNumDIs() + 0, var_OUT, conn_OUT);
-        break;
-      }
-      default: break;
-    }
-  }
-
-  CIEC_ANY *FORTE_SINT2SINT::getDI(const size_t paIndex) {
-    switch (paIndex) {
-      case 0: return &var_IN;
-    }
-    return nullptr;
-  }
-
-  CIEC_ANY *FORTE_SINT2SINT::getDO(const size_t paIndex) {
-    switch (paIndex) {
-      case 0: return &var_OUT;
-    }
-    return nullptr;
-  }
-
-  CEventConnection *FORTE_SINT2SINT::getEOConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_CNF;
-    }
-    return nullptr;
-  }
-
-  CDataConnection **FORTE_SINT2SINT::getDIConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_IN;
-    }
-    return nullptr;
-  }
-
-  CDataConnection *FORTE_SINT2SINT::getDOConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_OUT;
-    }
-    return nullptr;
-  }
-
-  CIEC_ANY *FORTE_SINT2SINT::getVarInternal(size_t) {
-    return nullptr;
-  }
-
-  void FORTE_SINT2SINT::alg_REQ(void) {
-
-#line 2 "SINT2SINT.fbt"
-    var_OUT = var_IN;
-  }
-
-} // namespace forte::eclipse4diac::convert
