--- modules/convert/src/eclipse4diac/convert/USINT2USINT_fbt.cpp
@@ -1,138 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2023 Martin Erich Jobst
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *   Martin Jobst
- *     - initial API and implementation and/or initial documentation
- *******************************************************************************/
-
-#include "forte/eclipse4diac/convert/USINT2USINT_fbt.h"
-
-#include "forte/datatypes/forte_usint.h"
-#include "forte/iec61131_functions.h"
-#include "forte/datatypes/forte_array_common.h"
-#include "forte/datatypes/forte_array.h"
-#include "forte/datatypes/forte_array_fixed.h"
-#include "forte/datatypes/forte_array_variable.h"
-
-using namespace forte::literals;
-
-namespace forte::eclipse4diac::convert {
-  namespace {
-    const auto cDataInputNames = std::array{"IN"_STRID};
-    const auto cDataOutputNames = std::array{"OUT"_STRID};
-    const auto cEventInputNames = std::array{"REQ"_STRID};
-    const auto cEventOutputNames = std::array{"CNF"_STRID};
-    const SFBInterfaceSpec cFBInterfaceSpec = {
-        .mEINames = cEventInputNames,
-        .mEITypeNames = {},
-        .mEONames = cEventOutputNames,
-        .mEOTypeNames = {},
-        .mDINames = cDataInputNames,
-        .mDONames = cDataOutputNames,
-        .mDIONames = {},
-        .mSocketNames = {},
-        .mPlugNames = {},
-    };
-  } // namespace
-
-  DEFINE_FIRMWARE_FB(FORTE_USINT2USINT, "eclipse4diac::convert::USINT2USINT"_STRID)
-
-  FORTE_USINT2USINT::FORTE_USINT2USINT(const StringId paInstanceNameId, CFBContainer &paContainer) :
-      CSimpleFB(paContainer, cFBInterfaceSpec, paInstanceNameId, {}),
-      var_IN(0_USINT),
-      var_OUT(0_USINT),
-      conn_CNF(*this, 0),
-      conn_IN(nullptr),
-      conn_OUT(*this, 0, var_OUT) {
-  }
-
-  void FORTE_USINT2USINT::setInitialValues() {
-    var_IN = 0_USINT;
-    var_OUT = 0_USINT;
-  }
-
-  void FORTE_USINT2USINT::executeEvent(const TEventID paEIID, CEventChainExecutionThread *const paECET) {
-    switch (paEIID) {
-      case scmEventREQID: enterStateREQ(paECET); break;
-      default: break;
-    }
-  }
-
-  void FORTE_USINT2USINT::enterStateREQ(CEventChainExecutionThread *const paECET) {
-    alg_REQ();
-    sendOutputEvent(scmEventCNFID, paECET);
-  }
-
-  void FORTE_USINT2USINT::readInputData(const TEventID paEIID) {
-    switch (paEIID) {
-      case scmEventREQID: {
-        readData(0, var_IN, conn_IN);
-        break;
-      }
-      default: break;
-    }
-  }
-
-  void FORTE_USINT2USINT::writeOutputData(const TEventID paEIID) {
-    switch (paEIID) {
-      case scmEventCNFID: {
-        writeData(cFBInterfaceSpec.getNumDIs() + 0, var_OUT, conn_OUT);
-        break;
-      }
-      default: break;
-    }
-  }
-
-  CIEC_ANY *FORTE_USINT2USINT::getDI(const size_t paIndex) {
-    switch (paIndex) {
-      case 0: return &var_IN;
-    }
-    return nullptr;
-  }
-
-  CIEC_ANY *FORTE_USINT2USINT::getDO(const size_t paIndex) {
-    switch (paIndex) {
-      case 0: return &var_OUT;
-    }
-    return nullptr;
-  }
-
-  CEventConnection *FORTE_USINT2USINT::getEOConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_CNF;
-    }
-    return nullptr;
-  }
-
-  CDataConnection **FORTE_USINT2USINT::getDIConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_IN;
-    }
-    return nullptr;
-  }
-
-  CDataConnection *FORTE_USINT2USINT::getDOConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_OUT;
-    }
-    return nullptr;
-  }
-
-  CIEC_ANY *FORTE_USINT2USINT::getVarInternal(size_t) {
-    return nullptr;
-  }
-
-  void FORTE_USINT2USINT::alg_REQ(void) {
-
-#line 2 "USINT2USINT.fbt"
-    var_OUT = var_IN;
-  }
-
-} // namespace forte::eclipse4diac::convert
