--- modules/convert/src/eclipse4diac/convert/WORD2WORD_fbt.cpp
@@ -1,138 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2023 Martin Erich Jobst
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *   Martin Jobst
- *     - initial API and implementation and/or initial documentation
- *******************************************************************************/
-
-#include "forte/eclipse4diac/convert/WORD2WORD_fbt.h"
-
-#include "forte/datatypes/forte_word.h"
-#include "forte/iec61131_functions.h"
-#include "forte/datatypes/forte_array_common.h"
-#include "forte/datatypes/forte_array.h"
-#include "forte/datatypes/forte_array_fixed.h"
-#include "forte/datatypes/forte_array_variable.h"
-
-using namespace forte::literals;
-
-namespace forte::eclipse4diac::convert {
-  namespace {
-    const auto cDataInputNames = std::array{"IN"_STRID};
-    const auto cDataOutputNames = std::array{"OUT"_STRID};
-    const auto cEventInputNames = std::array{"REQ"_STRID};
-    const auto cEventOutputNames = std::array{"CNF"_STRID};
-    const SFBInterfaceSpec cFBInterfaceSpec = {
-        .mEINames = cEventInputNames,
-        .mEITypeNames = {},
-        .mEONames = cEventOutputNames,
-        .mEOTypeNames = {},
-        .mDINames = cDataInputNames,
-        .mDONames = cDataOutputNames,
-        .mDIONames = {},
-        .mSocketNames = {},
-        .mPlugNames = {},
-    };
-  } // namespace
-
-  DEFINE_FIRMWARE_FB(FORTE_WORD2WORD, "eclipse4diac::convert::WORD2WORD"_STRID)
-
-  FORTE_WORD2WORD::FORTE_WORD2WORD(const StringId paInstanceNameId, CFBContainer &paContainer) :
-      CSimpleFB(paContainer, cFBInterfaceSpec, paInstanceNameId, {}),
-      var_IN(0_WORD),
-      var_OUT(0_WORD),
-      conn_CNF(*this, 0),
-      conn_IN(nullptr),
-      conn_OUT(*this, 0, var_OUT) {
-  }
-
-  void FORTE_WORD2WORD::setInitialValues() {
-    var_IN = 0_WORD;
-    var_OUT = 0_WORD;
-  }
-
-  void FORTE_WORD2WORD::executeEvent(const TEventID paEIID, CEventChainExecutionThread *const paECET) {
-    switch (paEIID) {
-      case scmEventREQID: enterStateREQ(paECET); break;
-      default: break;
-    }
-  }
-
-  void FORTE_WORD2WORD::enterStateREQ(CEventChainExecutionThread *const paECET) {
-    alg_REQ();
-    sendOutputEvent(scmEventCNFID, paECET);
-  }
-
-  void FORTE_WORD2WORD::readInputData(const TEventID paEIID) {
-    switch (paEIID) {
-      case scmEventREQID: {
-        readData(0, var_IN, conn_IN);
-        break;
-      }
-      default: break;
-    }
-  }
-
-  void FORTE_WORD2WORD::writeOutputData(const TEventID paEIID) {
-    switch (paEIID) {
-      case scmEventCNFID: {
-        writeData(cFBInterfaceSpec.getNumDIs() + 0, var_OUT, conn_OUT);
-        break;
-      }
-      default: break;
-    }
-  }
-
-  CIEC_ANY *FORTE_WORD2WORD::getDI(const size_t paIndex) {
-    switch (paIndex) {
-      case 0: return &var_IN;
-    }
-    return nullptr;
-  }
-
-  CIEC_ANY *FORTE_WORD2WORD::getDO(const size_t paIndex) {
-    switch (paIndex) {
-      case 0: return &var_OUT;
-    }
-    return nullptr;
-  }
-
-  CEventConnection *FORTE_WORD2WORD::getEOConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_CNF;
-    }
-    return nullptr;
-  }
-
-  CDataConnection **FORTE_WORD2WORD::getDIConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_IN;
-    }
-    return nullptr;
-  }
-
-  CDataConnection *FORTE_WORD2WORD::getDOConUnchecked(const TPortId paIndex) {
-    switch (paIndex) {
-      case 0: return &conn_OUT;
-    }
-    return nullptr;
-  }
-
-  CIEC_ANY *FORTE_WORD2WORD::getVarInternal(size_t) {
-    return nullptr;
-  }
-
-  void FORTE_WORD2WORD::alg_REQ(void) {
-
-#line 2 "WORD2WORD.fbt"
-    var_OUT = var_IN;
-  }
-
-} // namespace forte::eclipse4diac::convert
