+++ modules/utils/include/forte/eclipse4diac/utils/CMakeLists.txt
@@ -16,17 +16,10 @@ target_sources(forte-utils PUBLIC
         FILE_SET HEADERS
         FILES
         FB_RANDOM_fbt.h
-        GEN_ARRAY2ARRAY_fbt.h
-        GEN_ARRAY2VALUES_fbt.h
         GEN_CSV_WRITER_fbt.h
-        GEN_VALUES2ARRAY_fbt.h
-        GET_AT_INDEX_fbt.h
-        GET_STRUCT_VALUE_fbt.h
         GetInstancePath_fbt.h
         GetInstancePathAndName_fbt.h
         OUT_ANY_CONSOLE_fbt.h
-        SET_AT_INDEX_fbt.h
-        SET_STRUCT_VALUE_fbt.h
         STEST_END_fbt.h
         TEST_CONDITION_fbt.h
 )
